refactor: ATLAS-343: Split MP1 and MP2 logic into separate services and repositories.
