+++ Atlas.MatchingAlgorithm.Data/Repositories/DonorRetrieval/IPreFilteredDonorSearchRepository.cs
@@ -1,21 +1,20 @@
 ﻿using Atlas.Common.GeneticData;
 using Atlas.MatchingAlgorithm.Common.Models;
-using Atlas.MatchingAlgorithm.Common.Models.Matching;
 using Atlas.MatchingAlgorithm.Data.Models;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace Atlas.MatchingAlgorithm.Data.Repositories.DonorRetrieval
 {
-    public interface IDonorSearchRepository
+    public interface IPreFilteredDonorSearchRepository
     {
         /// <summary>
-        /// Returns donor matches at a given locus matching the search criteria
+        /// Returns donor matches at a given locus matching the search criteria, that are also present in a supplied list of donor ids
         /// </summary>
-        Task<IEnumerable<PotentialHlaMatchRelation>> GetDonorMatchesAtLocus(
+        Task<IEnumerable<PotentialHlaMatchRelation>> GetDonorMatchesAtLocusFromDonorSelection(
             Locus locus,
             LocusSearchCriteria criteria,
-            MatchingFilteringOptions filteringOptions
+            IEnumerable<int> donorIds
         );
     }
 }
